+++ src/core/foundations/src/mq/breakpoints.ts
@@ -1,0 +1,3 @@
+// This is a temporary fix to make migration work
+// as mq is not nested in packages/@guardian/source-foundations
+export { breakpoints, Breakpoint } from '../breakpoints';
