+++ src/core/foundations/src/mq/mq.ts
@@ -1,0 +1,1 @@
+../../../../../packages/@guardian/source-foundations/src/mq.ts
